restore original DAG with new pass task
